+++ tests/conftest.py
@@ -101,6 +101,7 @@ async def patch_enqueue(mocker, event_loop):
 
     def _execute(fn, *args, **kwargs):
         kwargs.pop("_priority")
+        kwargs.pop("_exception", None)
         result = fn(*args, **kwargs)
         if asyncio.iscoroutine(result):
             loop = event_loop
@@ -148,7 +149,11 @@ async def setup_catalog_with_resource_exception(setup_catalog):
     ['__id', 'Nom', 'Prenom', 'Societe', 'Adresse', 'CP', 'Ville', 'Tel1', 'Tel2', 'email', 'Organisme', 'Org Cofrac', 'Type de certificat', 'N° de certificat', 'Date début validité', 'Date fin validité']
     """
     await Resource.insert(
-        dataset_id=DATASET_ID, resource_id=RESOURCE_EXCEPTION_ID, url="http://example.com/"
+        dataset_id=DATASET_ID,
+        resource_id=RESOURCE_EXCEPTION_ID,
+        type="main",
+        format="csv",
+        url="http://example.com/",
     )
     await ResourceException.insert(
         resource_id=RESOURCE_EXCEPTION_ID,
@@ -194,6 +199,8 @@ async def insert_fake_resource():
             dataset_id=DATASET_ID,
             resource_id=RESOURCE_ID,
             url=RESOURCE_URL,
+            type="main",
+            format="csv",
             status=status,
             priority=True,
         )
@@ -280,6 +287,7 @@ def udata_resource_payload():
             "description": "random description",
             "filetype": "file",
             "type": "documentation",
+            "format": "pdf",
             "mime": "text/plain",
             "schema": None,
             "filesize": 1024,